+++ client/packages/invoices/src/Prescriptions/LineEditView/hooks/usePreviousNextItem.ts
