+++ client/packages/invoices/src/Prescriptions/LineEditView/index.ts
@@ -1,0 +1,1 @@
+export { PrescriptionLineEditView } from './LineEditView';
